feat(apollo,look&feel): ajouter le composant textarea
