+++ client/apollo/css/src/apollo.scss
@@ -5,6 +5,7 @@
 @use "./Form/ItemMessage/ItemMessageApollo";
 @use "./Form/ItemLabel/ItemLabelApollo";
 @use "./Form/TextInput/TextInputApollo";
+@use "./Form/TextArea/TextAreaApollo";
 @use "./Icon/IconApollo";
 @use "./Divider/DividerApollo";
 @use "./Link/LinkApollo";
