+++ client/apollo/css/src/common/tokens.scss
@@ -36,6 +36,7 @@
   --axa-blue-8: hsl(240deg, 33%, 91%);
   --axa-blue-80: hsl(240deg, 53%, 42%);
   --axa-blue-120: hsl(240deg, 100%, 22%);
+  --axa-blue-53: hsl(240deg, 39%, 53%);
 
   /* red */
 
@@ -44,6 +45,7 @@
   --axa-red-digital-100: hsl(357deg, 100%, 63%);
   --red-alert-100: hsl(0deg, 100%, 45%);
   --red-alert-80: hsl(0deg, 100%, 56%);
+  --red-alert-43: hsl(350deg, 82%, 43%);
 
   /* orange */
 
@@ -62,6 +64,11 @@
   --neutral-25: hsl(from var(--black) h s 80%);
   --neutral-14: hsl(from var(--black) h s 90%);
   --neutral-5: hsl(from var(--black) h s 96%);
+  --axa-gray-20: hsl(0deg, 0%, 20%);
+  --axa-gray-36: hsl(0deg, 0%, 36%);
+  --axa-gray-60: hsl(0deg, 0%, 60%);
+  --axa-gray-90: hsl(0deg, 0%, 90%);
+  --axa-gray-96: hsl(0deg, 0%, 96%);
 
   /* green */
   --green-100: hsl(138deg, 68%, 30%);
@@ -77,6 +84,8 @@
   --size-24: 24px;
   --size-32: 32px;
   --size-40: 40px;
+  --size-148: 148px;
+  --size-384: 384px;
 
   /* *************************************************************************/
