+++ client/apollo/css/src/look-and-feel.scss
@@ -5,6 +5,7 @@
 @use "./Form/ItemMessage/ItemMessageLF";
 @use "./Form/ItemLabel/ItemLabelLF";
 @use "./Form/TextInput/TextInputLF";
+@use "./Form/TextArea/TextAreaLF";
 @use "./Icon/IconLF";
 @use "./Divider/DividerLF";
 @use "./Link/LinkLF";
