+++ client/apollo/react/src/Form/TextArea/TextAreaApollo.tsx
@@ -1,0 +1,3 @@
+import "@axa-fr/design-system-apollo-css/dist/Form/TextArea/TextAreaApollo.scss";
+
+export { TextArea } from "./TextAreaCommon";
