+++ client/apollo/react/src/Form/TextArea/TextAreaLF.tsx
@@ -1,0 +1,3 @@
+import "@axa-fr/design-system-apollo-css/dist/Form/TextArea/TextAreaLF.scss";
+
+export { TextArea } from "./TextAreaCommon";
