+++ client/apollo/react/src/indexLF.ts
@@ -14,6 +14,7 @@ export { Tag, type TagVariants, tagVariants } from "./Tag/TagLF";
 export { DebugGrid } from "./Grid/DebugGrid";
 export { ItemMessage } from "./Form/ItemMessage/ItemMessageLF";
 export { ItemLabel } from "./Form/ItemLabel/ItemLabelLF";
+export { TextArea } from "./Form/TextArea/TextAreaLF";
 export { TextInput } from "./Form/TextInput/TextInputLF";
 export {
   Icon,
