--- client/look-and-feel/css/src/Form/TextArea/TextArea.scss
@@ -1,41 +1,0 @@
-@use "../../common/common" as common;
-
-.af-form__input-textarea {
-  width: 100%;
-  height: common.rem(142);
-  padding: 1rem 2.5rem 1rem 1rem;
-  border: 1px solid var(--color-gray);
-  border-radius: var(--default-border-radius);
-  font-size: common.rem(16);
-  line-height: common.rem(20);
-  color: var(--color-gray-900);
-  background-color: var(--color-white);
-
-  @media (width > #{common.$breakpoint-md}) {
-    font-size: common.rem(18);
-    line-height: common.rem(24);
-  }
-
-  &--error {
-    border: 1px solid var(--color-red-700);
-    box-shadow: 0 0 0 1px var(--color-red-700) inset;
-  }
-
-  &:not(:empty, :hover, :focus, :active, [class*="--error"]) {
-    border: 1px solid var(--color-axa);
-  }
-
-  &:enabled {
-    &:focus,
-    &:active {
-      border: 1px solid var(--color-axa);
-      outline: none;
-    }
-  }
-
-  &:not(:disabled, :focus, :active, &--error):hover {
-    border: 1px solid var(--color-axa);
-    box-shadow: 0 0 0 1px var(--color-axa) inset;
-    outline: none;
-  }
-}
