+++ client/look-and-feel/css/src/look-and-feel.scss
@@ -10,7 +10,6 @@
 @use "Tabs/Tabs";
 @use "Form/Checkbox/Checkbox";
 @use "Form/Select/Select";
-@use "Form/TextArea/TextArea";
 @use "Form/FileUpload/FileUpload";
 @use "Modal/Modal";
 @use "Pagination/Pagination";
