--- client/look-and-feel/react/src/Form/TextArea/index.ts
@@ -1,3 +1,0 @@
-import "@axa-fr/design-system-look-and-feel-css/dist/Form/TextArea/TextArea.scss";
-
-export { TextArea } from "./TextArea";
